uefi-services: deprecation: move to uefi::helpers
uefi_services::init temporarily stays API compatible to prevent
breaking changes. But we've added deprecation warnings.

The new uefi::helpers module consists mostly of an unchanged code import.
Only minor adjustments were applied. Cleanup will happen in a follow-up.
