+++ uefi-services/README.md
@@ -1,17 +1,3 @@
 # uefi-services
 
-[![Crates.io](https://img.shields.io/crates/v/uefi-services)](https://crates.io/crates/uefi-services)
-[![Docs.rs](https://docs.rs/uefi-macros/badge.svg)](https://docs.rs/uefi-services)
-
-This crate enables you some convenience features on top of the
-[`uefi`](https://crates.io/crates/uefi) crate. It includes a panic handler, a logger, and
-a global allocator.
-
-`uefi-services` is part of the `uefi-rs` project. Please refer to
-<https://github.com/rust-osdev/uefi-rs/> for comprehensive documentation.
-
-## Optional features
-
-This crate's features are described in [`src/lib.rs`].
-  
-[`src/lib.rs`]: src/lib.rs
+WARNING: `uefi-services` is deprecated. Functionality was moved to `uefi::helpers::init`.
